compiles without warnings, retabbed

+++ src/mapPreprocessor.cpp
@@ -69,9 +69,9 @@ vector<vector<float> > ratingsMap;
 
 typedef enum
 {
-	IDLE,
-	ACTIVE,
-	COMPLETE
+    IDLE,
+    ACTIVE,
+    COMPLETE
 }EMapLoaderState;
 EMapLoaderState state = IDLE;
 
@@ -79,245 +79,247 @@ EMapLoaderState state = IDLE;
    returns number of loaded maps  */
 int loadMaps()
 {
-	numMaps = 0;
-	DIR *dir;
-	struct dirent *ent;
-	cout << folder << endl;
-	if ((dir = opendir (folder.c_str())) != NULL) {
-		/* print all the files and directories within directory */
-		while ((ent = readdir (dir)) != NULL)
-		{
-			char filter[strlen(prefix.c_str())+10];
-			sprintf(filter,"%s_",prefix.c_str());
-			if (strstr(ent->d_name,"yaml") != NULL){
-				if (strncmp(ent->d_name,filter,strlen(filter)) == 0)  mapDistances[numMaps++] = atof(&ent->d_name[strlen(filter)]);
-			}
-		}
-		closedir (dir);
-	} else {
-		/* could not open directory */
-		ROS_ERROR("Could not open folder %s with maps.",folder.c_str());
-	}
+    numMaps = 0;
+    DIR *dir;
+    struct dirent *ent;
+    cout << folder << endl;
+    if ((dir = opendir (folder.c_str())) != NULL) {
+        /* print all the files and directories within directory */
+        while ((ent = readdir (dir)) != NULL)
+        {
+            char filter[strlen(prefix.c_str())+10];
+            sprintf(filter,"%s_",prefix.c_str());
+            if (strstr(ent->d_name,"yaml") != NULL){
+                if (strncmp(ent->d_name,filter,strlen(filter)) == 0)  mapDistances[numMaps++] = atof(&ent->d_name[strlen(filter)]);
+            }
+        }
+        closedir (dir);
+    } else {
+        /* could not open directory */
+        ROS_ERROR("Could not open folder %s with maps.",folder.c_str());
+    }
 
-	/* send feedback to action server */
-	feedback.fileName =  "";
-	feedback.numberOfMaps = numMaps;
-	//server->publishFeedback(feedback);
+    /* send feedback to action server */
+    feedback.fileName =  "";
+    feedback.numberOfMaps = numMaps;
+    //server->publishFeedback(feedback);
 
-	std::sort(mapDistances, mapDistances + numMaps, std::less<float>());
-	mapDistances[numMaps] = mapDistances[numMaps-1];
+    std::sort(mapDistances, mapDistances + numMaps, std::less<float>());
+    mapDistances[numMaps] = mapDistances[numMaps-1];
 
-	/*preload all maps*/
-	imagesMap.clear();
-	keypointsMap.clear();
-	descriptorMap.clear();
-	distanceMap.clear();
-	namesMap.clear();
-	ratingsMap.clear();
-	char fileName[1000];
+    /*preload all maps*/
+    imagesMap.clear();
+    keypointsMap.clear();
+    descriptorMap.clear();
+    distanceMap.clear();
+    namesMap.clear();
+    ratingsMap.clear();
+    char fileName[1000];
 
-	numFeatures=0;
-	for (int i = 0;i<numMaps;i++){
-		sprintf(fileName,"%s/%s_%.3f.yaml",folder.c_str(),prefix.c_str(),mapDistances[i]);
-		ROS_INFO("Preloading %s/%s_%.3f.yaml",folder.c_str(),prefix.c_str(),mapDistances[i]);
-		FileStorage fs(fileName, FileStorage::READ);
-		if(fs.isOpened())
-		{
-			img.release();
-			descriptors_1.release();
-			fs["Keypoints"]  >> keypoints_1;
-			fs["Descriptors"]>>descriptors_1;
-			fs["Image"]>>img;
-			ratings.clear();
-			fs["Ratings"]>>ratings;
-			for (int j = ratings.size(); j < keypoints_1.size(); j++) ratings.push_back(0);
-			fs.release();
-			keypointsMap.push_back(keypoints_1);
-			descriptorMap.push_back(descriptors_1);
-			distanceMap.push_back(mapDistances[i]);
-			namesMap.push_back(fileName);
-			ratingsMap.push_back(ratings);
-			if (image_pub_.getNumSubscribers()>0) imagesMap.push_back(img);
-			numFeatures+=keypoints_1.size();
-			sprintf(fileName,"Loading map %i/%i",i+1,numMaps);
-			feedback.fileName = fileName;
-			feedback.distance = mapDistances[i];
-			feedback.numFeatures=numFeatures;
-			feedback.mapIndex=i;
-			server->publishFeedback(feedback);
-		}
+    numFeatures=0;
+    for (int i = 0;i<numMaps;i++){
+        sprintf(fileName,"%s/%s_%.3f.yaml",folder.c_str(),prefix.c_str(),mapDistances[i]);
+        ROS_INFO("Preloading %s/%s_%.3f.yaml",folder.c_str(),prefix.c_str(),mapDistances[i]);
+        FileStorage fs(fileName, FileStorage::READ);
+        if(fs.isOpened())
+        {
+            img.release();
+            descriptors_1.release();
+            fs["Keypoints"]  >> keypoints_1;
+            fs["Descriptors"]>>descriptors_1;
+            fs["Image"]>>img;
+            ratings.clear();
+            fs["Ratings"]>>ratings;
+            for(size_t j = ratings.size(); j < keypoints_1.size(); j++) ratings.push_back(0);
+            fs.release();
+            keypointsMap.push_back(keypoints_1);
+            descriptorMap.push_back(descriptors_1);
+            distanceMap.push_back(mapDistances[i]);
+            namesMap.push_back(fileName);
+            ratingsMap.push_back(ratings);
+            if (image_pub_.getNumSubscribers()>0) imagesMap.push_back(img);
+            numFeatures+=keypoints_1.size();
+            sprintf(fileName,"Loading map %i/%i",i+1,numMaps);
+            feedback.fileName = fileName;
+            feedback.distance = mapDistances[i];
+            feedback.numFeatures=numFeatures;
+            feedback.mapIndex=i;
+            server->publishFeedback(feedback);
+        }
 
-	}
+    }
 
-	/* feedback returns name of loaded map, number of features in it and index */
-	sprintf(fileName,"%i features loaded from %i maps",numFeatures,numMaps);
-	feedback.fileName = fileName;
-	feedback.distance = mapDistances[numMaps-1];
-	feedback.numFeatures=numFeatures;
-	feedback.mapIndex=numMaps;
-	server->publishFeedback(feedback);
-	return numMaps;
+    /* feedback returns name of loaded map, number of features in it and index */
+    sprintf(fileName,"%i features loaded from %i maps",numFeatures,numMaps);
+    feedback.fileName = fileName;
+    feedback.distance = mapDistances[numMaps-1];
+    feedback.numFeatures=numFeatures;
+    feedback.mapIndex=numMaps;
+    server->publishFeedback(feedback);
+    return numMaps;
 }
 
 /* load map based on distance travelled  */
 void loadMap(int index)
 {
-	lastLoadedMap = index;
-	keypoints_1 = keypointsMap[index];
-	descriptors_1 = descriptorMap[index];
-	currentMapName = namesMap[index];
-	currentDistance = distanceMap[index];
-	ratings = ratingsMap[index];
-	numFeatures=keypoints_1.size();
-	char fileName[1000];
-	sprintf(fileName,"%i features loaded from %ith map at %.3f",numFeatures,index,distanceMap[index]);
-	feedback.fileName=fileName;
-	feedback.distance = currentDistance;
-	feedback.numFeatures=keypoints_1.size();
-	feedback.mapIndex=index;
-	/* feedback returns name of loaded map, number of features in it and index */
-	server->publishFeedback(feedback);
+    lastLoadedMap = index;
+    keypoints_1 = keypointsMap[index];
+    descriptors_1 = descriptorMap[index];
+    currentMapName = namesMap[index];
+    currentDistance = distanceMap[index];
+    ratings = ratingsMap[index];
+    numFeatures=keypoints_1.size();
+    char fileName[1000];
+    sprintf(fileName,"%i features loaded from %ith map at %.3f",numFeatures,index,distanceMap[index]);
+    feedback.fileName=fileName;
+    feedback.distance = currentDistance;
+    feedback.numFeatures=keypoints_1.size();
+    feedback.mapIndex=index;
+    /* feedback returns name of loaded map, number of features in it and index */
+    server->publishFeedback(feedback);
 }
 
 /* load path profile
    returns size of path profile */
 int loadPath()
 {
-	char fileName[1000];
-	sprintf(fileName,"%s/%s.yaml",folder.c_str(),prefix.c_str());
-	ROS_DEBUG("Loading %s/%s.yaml",folder.c_str(),prefix.c_str());
-	FileStorage fsp(fileName, FileStorage::READ);
-	vector<float> path;
-	path.clear();
-	if(fsp.isOpened()){
-		fsp["Path"]>>path;
-		fsp.release();
-	}
-	stroll_bearnav::PathProfile pathProfile;
-	for(int i=0;i<path.size()/4;i++){
-		pathProfile.distance.push_back(path[4*i+0]);
-		pathProfile.forwardSpeed.push_back(path[4*i+1]);
-		pathProfile.angularSpeed.push_back(path[4*i+2]);
-		pathProfile.flipper.push_back(path[4*i+3]);
-	}
-	pathPub.publish(pathProfile);
-	return path.size();
+    char fileName[1000];
+    sprintf(fileName,"%s/%s.yaml",folder.c_str(),prefix.c_str());
+    ROS_DEBUG("Loading %s/%s.yaml",folder.c_str(),prefix.c_str());
+    FileStorage fsp(fileName, FileStorage::READ);
+    vector<float> path;
+    path.clear();
+    if(fsp.isOpened()){
+        fsp["Path"]>>path;
+        fsp.release();
+    }
+    stroll_bearnav::PathProfile pathProfile;
+    for(size_t i = 0; i < path.size()/4; i++)
+    {
+        pathProfile.distance.push_back(path[4*i+0]);
+        pathProfile.forwardSpeed.push_back(path[4*i+1]);
+        pathProfile.angularSpeed.push_back(path[4*i+2]);
+        pathProfile.flipper.push_back(path[4*i+3]);
+    }
+    pathPub.publish(pathProfile);
+    return path.size();
 }
 
 /* Action server */
 void executeCB(const stroll_bearnav::loadMapGoalConstPtr &goal, Server *serv)
 {
-	lastLoadedMap = -1;
-	numProcessedMaps = 0;
-	stroll_bearnav::loadMapFeedback feedback;
+    lastLoadedMap = -1;
+    numProcessedMaps = 0;
+    stroll_bearnav::loadMapFeedback feedback;
 
-	prefix = goal->prefix;
+    prefix = goal->prefix;
 
-	if (loadMaps() > 0 && loadPath() >= 0){
-		state = ACTIVE;
-	}else{
-		result.distance=0;
-		result.numFeatures=0;
-		result.numMaps=0;
-		server->setAborted(result);
-	}
-	/* server returns distance travelled, total number of features
-	   and number of loaded map on preempt request*/
-	while(state != IDLE){
-		usleep(200000);
+    if (loadMaps() > 0 && loadPath() >= 0){
+        state = ACTIVE;
+    }else{
+        result.distance=0;
+        result.numFeatures=0;
+        result.numMaps=0;
+        server->setAborted(result);
+    }
+    /* server returns distance travelled, total number of features
+       and number of loaded map on preempt request*/
+    while(state != IDLE){
+        usleep(200000);
 
-		if(server->isPreemptRequested()){
-			result.distance=distanceT;
-			result.numFeatures=numFeatures;
-			result.numMaps=numProcessedMaps;
-			server->setPreempted(result);
-			state = IDLE;
-		}
-		if(state == COMPLETE){
-			result.distance=distanceT;
-			result.numFeatures=numFeatures;
-			result.numMaps=numProcessedMaps;
-			server->setSucceeded(result);
-			state = IDLE;
-		}
-	}
+        if(server->isPreemptRequested()){
+            result.distance=distanceT;
+            result.numFeatures=numFeatures;
+            result.numMaps=numProcessedMaps;
+            server->setPreempted(result);
+            state = IDLE;
+        }
+        if(state == COMPLETE){
+            result.distance=distanceT;
+            result.numFeatures=numFeatures;
+            result.numMaps=numProcessedMaps;
+            server->setSucceeded(result);
+            state = IDLE;
+        }
+    }
 }
 
 void distCallback(const std_msgs::Float32::ConstPtr& msg)
 {
-	if(state == ACTIVE){
-		distanceT=msg->data;
-		featureArray.feature.clear();
+    if(state == ACTIVE){
+        distanceT=msg->data;
+        featureArray.feature.clear();
 
-		//find the closest map
-		int mindex = -1;
-		float minDistance = FLT_MAX;
-		for (int i = 0;i<numMaps;i++){
-			if (fabs(distanceT-mapDistances[i]) < minDistance){
-				minDistance = fabs(distanceT-mapDistances[i]);
-				mindex = i;
-			}
-		}
+        //find the closest map
+        int mindex = -1;
+        float minDistance = FLT_MAX;
+        for (int i = 0;i<numMaps;i++){
+            if (fabs(distanceT-mapDistances[i]) < minDistance){
+                minDistance = fabs(distanceT-mapDistances[i]);
+                mindex = i;
+            }
+        }
 
-		//and publish it
-		if (mindex > -1 && mindex != lastLoadedMap){
-			ROS_INFO("Current distance is %.3f Closest map found at %i, last was %i",distanceT,mindex,lastLoadedMap);
-			loadMap(mindex);
-			//			ROS_INFO("Sending a map %i features with %i descriptors",(int)keypoints_1.size(),descriptors_1.rows);
+        //and publish it
+        if (mindex > -1 && mindex != lastLoadedMap){
+            ROS_INFO("Current distance is %.3f Closest map found at %i, last was %i",distanceT,mindex,lastLoadedMap);
+            loadMap(mindex);
+            //          ROS_INFO("Sending a map %i features with %i descriptors",(int)keypoints_1.size(),descriptors_1.rows);
 
-			int stcs[keypoints_1.size()];
-			for(int i = 0; i<keypoints_1.size();i++){
-				stcs[i] = 0;
-			}
+            int stcs[keypoints_1.size()];
+            for(size_t i = 0; i < keypoints_1.size(); i++)
+            {
+                stcs[i] = 0;
+            }
 
-			for(int i=0;i<keypoints_1.size();i++)
-			{
-				if(stcs[i]>=0)//(max/2)
-				{
-					feature.x=keypoints_1[i].pt.x;
-					feature.y=keypoints_1[i].pt.y;
-					feature.size=keypoints_1[i].size;
-					feature.angle=keypoints_1[i].angle;
-					feature.response=keypoints_1[i].response;
-					feature.octave=keypoints_1[i].octave;
-					feature.class_id=keypoints_1[i].class_id;
-					feature.descriptor=descriptors_1.row(i);
-					feature.rating=ratings[i];
-					featureArray.feature.push_back(feature);
-				}
-			}
-			featureArray.distance = currentDistance;
-			featureArray.id = currentMapName;
-			numberOfUsedMaps++;
+            for(size_t i = 0; i < keypoints_1.size(); i++)
+            {
+                if(stcs[i]>=0)//(max/2)
+                {
+                    feature.x=keypoints_1[i].pt.x;
+                    feature.y=keypoints_1[i].pt.y;
+                    feature.size=keypoints_1[i].size;
+                    feature.angle=keypoints_1[i].angle;
+                    feature.response=keypoints_1[i].response;
+                    feature.octave=keypoints_1[i].octave;
+                    feature.class_id=keypoints_1[i].class_id;
+                    feature.descriptor=descriptors_1.row(i);
+                    feature.rating=ratings[i];
+                    featureArray.feature.push_back(feature);
+                }
+            }
+            featureArray.distance = currentDistance;
+            featureArray.id = currentMapName;
+            numberOfUsedMaps++;
 
-			/* publish loaded map */
-			feat_pub_.publish(featureArray);
+            /* publish loaded map */
+            feat_pub_.publish(featureArray);
 
-			/*if someone listens, then publish loaded image too*/
-			if (image_pub_.getNumSubscribers()>0){
-				std_msgs::Header header;
-				cv_bridge::CvImage bridge(header, sensor_msgs::image_encodings::MONO8, imagesMap[mindex]);
-				image_pub_.publish(bridge.toImageMsg());
-			}
-		}
-	}
+            /*if someone listens, then publish loaded image too*/
+            if (image_pub_.getNumSubscribers()>0){
+                std_msgs::Header header;
+                cv_bridge::CvImage bridge(header, sensor_msgs::image_encodings::MONO8, imagesMap[mindex]);
+                image_pub_.publish(bridge.toImageMsg());
+            }
+        }
+    }
 }
 
 int main(int argc, char** argv)
 {
-	ros::init(argc, argv, "map_preprocessor");
-	ros::NodeHandle nh_;
-	image_transport::ImageTransport it_(nh_);
-	ros::param::get("~folder", folder);
-	cmd_pub_ = nh_.advertise<geometry_msgs::Twist>("/cmd",1);
-	pathPub = nh_.advertise<stroll_bearnav::PathProfile>("pathProfile",1);
-	dist_sub_ = nh_.subscribe<std_msgs::Float32>( "distance", 1,distCallback);
-	image_pub_ = it_.advertise("map_image", 1);
-	feat_pub_ = nh_.advertise<stroll_bearnav::FeatureArray>("localMap",1);
+    ros::init(argc, argv, "map_preprocessor");
+    ros::NodeHandle nh_;
+    image_transport::ImageTransport it_(nh_);
+    ros::param::get("~folder", folder);
+    cmd_pub_ = nh_.advertise<geometry_msgs::Twist>("/cmd",1);
+    pathPub = nh_.advertise<stroll_bearnav::PathProfile>("pathProfile",1);
+    dist_sub_ = nh_.subscribe<std_msgs::Float32>( "distance", 1,distCallback);
+    image_pub_ = it_.advertise("map_image", 1);
+    feat_pub_ = nh_.advertise<stroll_bearnav::FeatureArray>("localMap",1);
 
-	/* Initiate action server */
-	server = new Server (nh_, "map_preprocessor", boost::bind(&executeCB, _1, server), false);
-	server->start();
-	ros::spin();
-	return 0;
+    /* Initiate action server */
+    server = new Server (nh_, "map_preprocessor", boost::bind(&executeCB, _1, server), false);
+    server->start();
+    ros::spin();
+    return 0;
 }
